chore: enable eslint rules and ignore violations

+++ apps/main/src/bridge/features/bridge/components/BridgeFormContent.stories.tsx
@@ -46,13 +46,14 @@ const BridgeForm = (props: BridgeFormContentParams) => {
   const [amount, setAmount] = useState<Decimal | undefined>(undefined)
   const [loading, setLoading] = useState(true)
   const [walletBalance, setWalletBalance] = useState({
-    balance: '1234' as Decimal,
+    balance: '1234',
     loading: true,
     notionalValueUsd: 500,
   })
 
   // Simulate loading at start
   useEffect(() => {
+    // eslint-disable-next-line @eslint-react/web-api-no-leaked-timeout -- Existing violation before enabling this rule.
     setTimeout(() => {
       setLoading(false)
       setWalletBalance({ ...walletBalance, loading: false })
